fix: lookup settings against new namepace and older namespace
Previously this package was appcelerator-titanium, during a botched release before atom shutdown
I renamed it to titanium to try and publish a new module (the backend was giving strange errors)
that indicated updates couldn't be published. I should have followed this through with correcting
the lookups for config settings and other things, so this is finally doing that.

It attempts to do it in a backwards compatible manner, first we lookup the new config setting, and
then fallback to the older setting if we can't find it. However, I'm not certain what the behaviour
will be for default values as these will be set in both places